addcourse edited

+++ Interlearn/app/controllers/receptionist.php
@@ -15,153 +15,74 @@ class Receptionist extends Controller
     }
 
     public function course($action = null, $id = null)
-    {
+    { 
         if(!Auth::is_receptionist()){
             redirect('home');
-
+           
         }
+
         $user_id = Auth::getuid();
         $subject = new Subject();
         $course = new Course();
         $teacher = new Teacher();
         $instructor = new Instructor();
+        $course_week = new CourseWeek();
+        // $instructor_course = new InstructorCourse();
         $data = [];
         // echo $user_id;die;
         // $teacher_id = $teacher-> selectTeacher(['uid'=>$user_id]);
-
+       
         if($action == 'add'){
-
+   
             $data = [];
             $data['action'] = $action;
-            $data['id'] = $id;
+            $data['id'] = $id; 
 
-            // $data['days'] = $day->findAll('asc');
+
+            $data['subjects'] = $subject->getSubject();
+            $data['grades'] = $subject->getGrade();
             $data['teachers'] = $teacher->select([],'teacher_id','asc');
             //show($data['teachers']);die;
             $data['instructors'] = $instructor->select([],'instructor_id','asc');
             //show($data['instructors']);die;
-
-            // if(isset($_REQUEST['courseSubmitBtn']))
+        
+            if ($_SERVER['REQUEST_METHOD'] == 'POST') 
+             {   
+            //     if(($_POST['subject'] == "") || ($_POST['language_medium'] == "") || ($_POST['grade'] == "") || ($_POST['teacher_id'] == "") || ($_POST['day'] == "") || ($_POST['timefrom'] == "") || ($_POST['timeto'] == ""))
             //     {
-            //         //checking for empty fields
-            //         if(($_REQUEST['subject'] == "") || ($_REQUEST['description'] == "") || ($_REQUEST['language_medium'] == "") || ($_REQUEST['grade'] == "") || ($_REQUEST['teacher_id'] == "") || ($_REQUEST['day'] == "") || ($_REQUEST['timefrom'] == "") || ($_REQUEST['timeto'] == "") ||($_REQUEST['uploadimg'] == ""))
-            //         {
-            //             $msg = '<div class="alert">Fill all the fields!</div>';
-            //         }
+            //         $msg = '<div class="alert">Fill all the fields!</div>';
+            //     }else{
+            //         $subject_name = $_REQUEST['subject'];
+            //         $language_medium = $_REQUEST['language_medium'] ;
+            //         $grade = $_REQUEST['grade'];
+            //         $teacher_id = $_REQUEST['teacher_id'];
+            //         $day = $_REQUEST['day'];
+            //         $timefrom = $_REQUEST['timefrom'];
+            //         $timeto = $_REQUEST['timeto'];
             //     }
 
-            if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-// var_dump($_POST);
-// exit;
-                if(($_POST['subject'] == "") || ($_POST['description'] == "") || ($_POST['language_medium'] == "") || ($_POST['grade'] == "") || ($_POST['teacher_id'] == "") || ($_POST['day'] == "") || ($_POST['timefrom'] == "") || ($_POST['timeto'] == "") ||($_POST['uploadimg'] == ""))
-                {
-                    $msg = '<div class="alert">Fill all the fields!</div>';
-                }else{
-                    $subject = $_REQUEST['subject'];
-                    $description = $_REQUEST['description'] ;
-                    $language_medium = $_REQUEST['language_medium'] ;
-                    $grade = $_REQUEST['grade'];
-                    $teacher_id = $_REQUEST['teacher_id'];
-                    $day = $_REQUEST['day'];
-                    $timefrom = $_REQUEST['timefrom'];
-                    $timeto = $_REQUEST['timeto'];
-                    $uploadimg = $_FILES['uploadimg']['name'];
-                }
+                
 
-                if (isset($_POST["submit"]))
-                {
-                    $target_dir = "uploads/images/";
-                    $target_file = $target_dir . basename($_FILES["uploadimg"]["name"]);
-                    $imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
+                $subject_id = uniqid();
+                //uniqueid("S",true)
 
-                    // Check if image file is a actual image or fake image
-                    $check = getimagesize($_FILES["uploadimg"]["tmp_name"]);
-                    if($check !== false) {
-                      // Check file size
-                      if ($_FILES["uploadimg"]["size"] > 500000) {
-                        echo "Sorry, your file is too large.";
-                      } else {
-                        // Allow certain file formats
-                        if($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg"
-                        && $imageFileType != "gif" ) {
-                          echo "Sorry, only JPG, JPEG, PNG & GIF files are allowed.";
-                        } else {
-                          // Upload file
-                          if (move_uploaded_file($_FILES["uploadimg"]["tmp_name"], $target_file)) {
-                            // var_dump($_FILES);die;
-                            // $course->insert($_FILES["uploadimg"]["name"]);
-                            echo "The file ". htmlspecialchars( basename( $_FILES["uploadimg"]["name"])). " has been uploaded.";
+                 $_POST['subject_id']=$subject_id;
+                $subject->insert($_POST);
+                
+                // $subject_details = $subject->getSubjectId($_POST['subject'],$_POST['grade'],$_POST['language_medium']);
 
-                          }else {
-                            echo "Sorry, there was an error uploading your file.";
-                          }
-                        }
-                      }
-                    } else {
-                      echo "File is not an image.";
-                    }
-                }
+                // $_POST['subject_id'] = $subject_details[0]->subject_id;
+                // print_r ($subject_id);die;
+                $course->insert($_POST);
 
+                // // show($subject_id);die;
+                $id= $course->getLastCourse()[0]->course_id;
+                // // // print_r($Course);die;
+                $course_week->createWeek($id, 1);
 
-
-                if(isset($_POST['subject']))
-                {
-                    $subject_id = uniqid();
-                    $_POST['subject_id']=$subject_id;
-                    $result = $subject->insert($_POST);
-                }
-                // $_POST['subject_id']='1';
-                $result2 = $course->insert($_POST);
-                    // show ($course);die;
-                    // $row = $course->first(['$teacher_id'=>$teacher_id]);
-                    // $sum = $subject->first(['subject_id'=>$subject_id],'subject_id');
-
-                    // message("Your course was successfully created.");
-                    // if($row){
-                    //     redirect('receptionist/course/edit/'.$row->id);
-                    // }else{
-                    //     redirect('receptionist/course');
-                    // }
-
-                    if($result == TRUE && $result2 == TRUE){
-                        $msg = '<div class="alert">Course added successfully!</div>';
-                    }else{
-                        $msg = '<div class="alert">Unable to add course!</div>';
-                    }
             }
             $this->view('receptionist/addCourse',$data);
-            exit;
-
-            // if(isset($_FILES['uploadimg']['name']) AND !empty($_FILES['uploadimg']['name'])){
-            //     $pic_tmp = $_FILES['uploadimg']["tmp_name"];
-            //     $pic_name = $_FILES['uploadimg']["name"];
-            //     $error= $_FILES['uploadimg']['error'];
-
-            //     if($error === 0){
-            //         $img_ext = pathinfo($pic_name,PATHINFO_EXTENSION);
-            //         $img_final_ext = strtolower($img_ext);
-
-            //         $allowed_ext = array('jpg','png','jpeg');
-            //         if(in_array($img_final_ext,$allowed_ext)){
-            //             $new_image_name = uniqid($_POST['username'],true).'.'.$img_final_ext;
-            //             $destination = "uploads/images/". $new_image_name;
-            //           // echo $destination;die;
-            //             move_uploaded_file($pic_tmp,$destination);
-            //             $_POST['display_picture'] = $new_image_name ;
-            //         }
-            //         else{
-            //             $data['errors']['uploadimg']='you cannot upload this type of file';
-            //         }
-            //     }
-            //     else{
-            //         $data['errors']['uploadimg'] ="unknown error occured";
-            //     }
-            // }
-
-
-
-
-
+            exit;   
         }
 
         if($action == 'view')
@@ -169,14 +90,16 @@ class Receptionist extends Controller
             if(isset($_POST['save-btn']))
             {
                 $inputs=array("subject_id"=>$_GET['id'],"teacher_id"=>$_POST['teacher_id'],"day"=>$_POST['day'],"timefrom"=>$_POST['timefrom'],"timeto"=>$_POST['timeto']);
+                //show($inputs);die;
                 $course->insert($inputs);
+                //show($_POST);die;
             }
             if(isset($_POST['save-btn2']))
             {
                 $inputs=array("subject_id"=>$_GET['id'],"instructor_id"=>$_POST['instructor_id'],"day"=>$_POST['day'],"timefrom"=>$_POST['timefrom'],"timeto"=>$_POST['timeto']);
                 $course->insert($inputs);
             }
-
+            
                 $data = [];
                 $data['action'] = $action;
                 $data['id'] = $id;
@@ -187,32 +110,47 @@ class Receptionist extends Controller
                     if(isset($_GET['id'])){
                         $subject_id = $_GET['id'];
                         $data['subject_id'] = $subject_id;
-                        $allSubjects = $subject -> where(['subject_id'=>$subject_id],'subject_id');
-                        $data['subjectgrd'] = $allSubjects;
+                        // $allSubjects = $subject -> where(['subject_id'=>$subject_id],'subject_id');
+                        $subjectGrade = $subject -> getSubjectGrade($subject_id);
+                        // show($subjectGrade->subject_id);die;
+                        $subjectName = $subjectGrade[0]->subject;
+                        $grade = $subjectGrade[0]->grade;
+                        //  echo $subjectName;
+                        //  echo $grade;die;
+                        $data['mediums'] = $subject -> getMedium($subjectName,$grade);
+                          //show($data['mediums']);die;
+                         
                         //show($data['subjectgrd']);die;
-                                 //show($allSubjects);die;
+                                // show($allSubjects);die;
+                               
                     $medium = "Sinhala";
-
-                    $data['subjects']=$subject->selectTeachers(['subject'=>$allSubjects[0]->subject, 'grade'=>$allSubjects[0]->grade],$medium,$subject_id);
-                    if($id==1){
-                        $medium = "Sinhala";
-                        $data['subjects']=$subject->selectTeachers(['subject'=>$allSubjects[0]->subject, 'grade'=>$allSubjects[0]->grade],$medium,$subject_id);
-
-                        //show($data['subjects']);die;
-
-                    }
-                    if($id==2){
-                        $medium = "English";
-                        $data['subjects']=$subject->selectTeachers(['subject'=>$allSubjects[0]->subject, 'grade'=>$allSubjects[0]->grade],$medium,$subject_id);
-                    }
-                    if($id==3){
-                        $medium = "Tamil";
-                        $data['subjects']=$subject->selectTeachers(['subject'=>$allSubjects[0]->subject, 'grade'=>$allSubjects[0]->grade],$medium,$subject_id);
-
-
-                        //show($data['subjects']);die;
-
-                    }
+                    
+                    //  $data['subjects']=$subject->selectTeachers(['subject'=>$allSubjects[0]->subject, 'grade'=>$allSubjects[0]->grade],$medium,$subject_id);
+                     $data['subjects'] = $subject -> selectTeachers(['subject_id'=>$data['mediums'][0]->subject_id],$data['mediums'][0]->language_medium);
+                    
+                    $data['sinhalaid'] = $subject->getSubjectId($subjectName,$grade,"Sinhala");
+                    $data['englishid'] = $subject->getSubjectId($subjectName,$grade,"English");
+                    $data['tamilid'] = $subject->getSubjectId($subjectName,$grade,"Tamil");
+                    //show($data['subjects']);die;
+                    // if($id==1){
+                    //     $medium = "Sinhala";
+                    //     $data['subjects']=$subject->selectTeachers(['subject'=>$allSubjects[0]->subject, 'grade'=>$allSubjects[0]->grade],$medium,$subject_id);
+                        
+                    //     //show($data['subjects']);die;
+                        
+                    // }
+                    // if($id==2){
+                    //     $medium = "English";
+                    //     $data['subjects']=$subject->selectTeachers(['subject'=>$allSubjects[0]->subject, 'grade'=>$allSubjects[0]->grade],$medium,$subject_id);
+                    // }
+                    // if($id==3){
+                    //     $medium = "Tamil";
+                    //     $data['subjects']=$subject->selectTeachers(['subject'=>$allSubjects[0]->subject, 'grade'=>$allSubjects[0]->grade],$medium,$subject_id);
+                     
+                        
+                    //     //show($data['subjects']);die;
+                    
+                    // }
                 }
                 $data['teachers'] = $teacher->select([],'teacher_id','asc');
                 $data['instructors'] = $instructor->select([],'instructor_id','asc');
@@ -221,35 +159,27 @@ class Receptionist extends Controller
                 exit;
         }
 
-        // if($action == 'edit'){
-        //     if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-        //         $announcement_id = uniqid();
-        //         $_POST['aid'] = $announcement_id;
-        //         //show($_POST);die;
-        //         //$announcement = new Announcement();
-        //         // $result = $announcement->insert($_POST);
-        //         // $result2 = $ann_course->insert($_POST);
-        //         echo "Announcement successfully published!";
-        //         //show($_POST);die;
-
-        //     }
-        //     $this->view('receptionist/class',$data);
-        // }
+       
 
         if($action == 'delete'){
-            $result = $course->delete(['course_id'=>$id]);
-            header("Location:http://localhost/Interlearn/public/receptionist/course");
+            if(isset($_POST['submit-delete-course'])){
+                $result = $course->delete(['course_id'=>$id]);
+                header("Location:http://localhost/Interlearn/public/receptionist/course");
+            }
+            // $data = [];
+            // $data['action'] = $action;
+            // $data['id'] = $id;
+            
             $this->view('receptionist/class',$data);
         }
-
+       
         $data['rows']= $course->select([],'course_id');
         $data['sums']= $subject -> distinctSubject([],'subject');
-        //  show($data['sums']);die;
-
-
+          //show($data['sums']);die;
+       
+           
          $this->view('receptionist/course',$data);
     }
-
     public function class($action=null, $id = null)
     {
         if(!Auth::is_receptionist()){
@@ -476,36 +406,38 @@ class Receptionist extends Controller
             
         $this->view('receptionist/enquiry',$data);
     }
-}
 
-public function payments()
-{
-    if (!Auth::is_receptionist()) {
-        redirect('home');
+
+    public function payments()
+    {
+        if (!Auth::is_receptionist()) {
+            redirect('home');
+        }
+
+        $this->view('receptionist/receptionist-payments');
     }
 
-    $this->view('receptionist/receptionist-payments');
-}
-
-public function getPayment()
-{
-    if (!Auth::is_receptionist()) {
-        redirect('home');
+    public function getPayment()
+    {
+        if (!Auth::is_receptionist()) {
+            redirect('home');
+        }
+        // show($_POST);
+        if (isset($_POST)) {
+            $payment_model = new Payment();
+            $_POST['method'] = 'cash';
+            $_POST['status'] = '1';
+            $payment_model->insert($_POST);
+            $this->view("receptionist/receptionist-payments");
+        }
     }
-    // show($_POST);
-    if (isset($_POST)) {
-        $payment_model = new Payment();
-        $_POST['method'] = 'cash';
-        $_POST['status'] = '1';
-        $payment_model->insert($_POST);
-        $this->view("receptionist/receptionist-payments");
+
+
+    public function getPaymentData()
+    {
+        $payment = new Payment();
+        $data = $payment->getAll();
+        return $data;
     }
-}
 
-
-public function getPaymentData()
-{
-    $payment = new Payment();
-    $data = $payment->getAll();
-    return $data;
 }
